Replace wx sizers with custom layouts
This removes all traces of wx's sizers except in bash.py.
They are replaced with layouts, a group of classes that
should be abstracted enough to work with not only wx but also other gui
libraries if needed.

The design of the layout API is a mix between wx, qt, and the existing
thin wrappers around the sizers in bash. They are not quite as flexible
as wx's sizers yet, but they should be easier to understand and less
error prone. It's still a WIP however so suggestions for improvements
are very welcome.

Visually, most windows and widgets should not be very different as I
tried to stay with the original design, but in some places there might
be some differences (especially when it comes to spacing and borders).

Under # 190