Rework backup and restore settings functions:
The backup and restore settings functions where changed to only take the
actually relevant variables as arguments, not calculate should_quit
booleans inside the function. The decisions if a backup/restore is
executed was moved outside of the functions, so their name now describes
their function. Also added additional logging and error reporting.